Set format's mountpoint if it has the mountpoint attribute.
Checking whether the device is mountable is somewhat expensive. Also, it
seems like the value of mountable could change. It would be depressing to
have not set the mountpoint here, but to later encounter the mountable() but
alas, no mountpoint, situation.

I can not find a location where the existance of a mountpoint is taken
to indicate that the format is mountable.

Signed-off-by: mulhern <amulhern@redhat.com>