Make implicit partitions smaller if real requests don't fit anywhere
In case of BTRFS/LVM autopartitioning we schedule one parition on every disk
used for the installation to be part of the new BTRFS/LVM setup. But in case
some other partition needs to be allocated (e.g. swap in case of BTRFS) we need
to fit this partition to some disk together with one of the implicitly scheduled
ones. In case we find out the implicitly scheduled partitions are too big to
make this possible we need to make them smaller so that autopartitioning works.

Resolves: rhbz#1171116
Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>