Use uevents to determine when device actions are complete.
Add class threads.StorageSynchronizer, which facilitates communication
between action threads and uevent handler threads to coordinate the
finalization of actions. StorageSynchronizer uses a condition variable
and a set of flags that indicate what type of operation is underway.
Every device and format instance contains an instance of
StorageSynchronizer. All instances of StorageSynchronizer use the
reentrant global blivet lock as the underlying lock for the condition
variable. When the underlying operation (eg: mkfs, lvcreate, wipefs)
that implements an action is finished, the device or format waits for
notification from the uevent handler to tell it the appropriate uevent
has been received. Then the handler waits for the device or format to
do any post-processing, like setting self.exists to True (for create)
or False (for destroy). Once the device or format has completed post-
processing, it notifies the uevent handler that it can finish.

A somewhat simpler explanation, in which there are two threads:
the action thread, which is executing a DeviceAction, and the event
thread, which is spawned to handle a uevent:

1. [action] device/format performs operation, which generates a uevent
2. [action] device/format waits until notified of receipt of uevent
3. [event] uevent handler receives event, associates with device/format
4. [event] uevent handler notifies device/format of uevent receipt
5. [action] device/format does post-processing
6. [action] device/format notifies uevent handler that post-processing
            is complete
7. [event] uevent handler continues with normal handling of uevent

Partition create and destroy are surprisingly complex and necessitated
an additional wait/notify cycle in the above model. Due to the number
of uevents generated by partition add/remove, it often happened that
the uevent handler would notify the device it had received the event
before the device was waiting for the notification, leading to a
deadlock. For this reason, when handling events from device create and
destroy operations, the handler first waits for the device to notify it
that the device is ready to be notified of the event:

1. [action] device performs operation, which generates a uevent
?.1. [action] device notifies uevent handler that it is ready to be
            notified of uevent receipt
?.2. [action] device waits until notified of receipt of uevent
?.1. [event] uevent handler receives uevent, associates with device
?.2. [event] uevent handler waits until device is ready to be notified
           of uevent receipt
6. [event] uevent handler notifies device of uevent receipt
7. [action] device does post-processing
8. [action] device notifies uevent handler that post-processing is
            complete
9. [event] uevent handler continues with normal handling of uevent

Fun, right?

TODO:
- handle devices not managed by udev (file, nodev, btrfs)
- handle devices that don't get automatically activated upon creation
  (lvn thin snapshots?)