Add the ability to blacklist certain device/action pairs.
It is possible to blacklist based on device, action, or both. It is
also possible to set a blacklist such that it only ignores a fixed
number of matching events before resuming normal operation.

This is handy for disklabels. When we get a change event on a device that
contains a disklabel we have to delete the existing parted.Disk and create
a new one to pick up any changes. One of those operations causes another
change uevent to get generated, potentially leading to an infinite loop.
With this patch we can simply set a one-time blacklist on change events on
the disk right before we refresh the parted.Disk.