Set TmpFSDevice object's _formatImmutable attribute to True.
Otherwise, in blivet.Blivet.createDevice(), an action to create the
device's format will be scheduled. Scheduling the DeviceFormatCreate action
will raise an exception, since every TmpFS format object is defined to
always exist.

With this change no exception is raised in DeviceFormatCreate action
initializer, and the partition with fstype specified as tmpfs in kickstart
is created during installation with filesystem type tmpfs.

Signed-off-by: mulhern <amulhern@redhat.com>