Adapt better to libblockdev's md_examine data
We no longer need the try-except block because what we get is an object that has
all the fields we expect defined.

Also don't use md_info.name since that's mdadm's MD_NAME (e.g. localhost:pv00)
not MD_DEVNAME (pv00) and it is thus safer to construct the name from the device
path.