Simplify and correct StorageDevice.packages property.
It's an error to check the parent's format's packages in the loop,
just because it's possible to check self's format's packages.

The for statement was a null op, because all the packages that it considered
had already been added by the magic of virtual method invocation.
This would not be the case if any device classes extended Device w/out
extending StorageDevice...in that case there might have been an
AttributeError due to a missing format attribuate.

packages is not supposed to have duplicates, so make sure not to add
duplicates when adding format's packages.

So, StorageDevice.packages might previously have had duplicates, but now
it should not, as was clearly originally intended.

Update Device.packages so that method docstring is correct and simplify
body slightly.

Signed-off-by: mulhern <amulhern@redhat.com>