Add scratch, scratch-bumpver and rc-release targets.
Use these make targets for creating scratch rc builds that can be used
before an official release. rc-release will run scratch-bumpver to bump
the version and add a rc release with the current date and time to the
specfile.  scratch is then run to create the tar.gz file and then mock
is used to build a rc rpm package.

No changes are committed to git, no zanata account is required and
nothing is pushed to zanata.