Add support for tracking external dependencies of devices.
Related: #12

Every device type can store in _external_dependencies the resources that its
code obviously depends on and the methods can traverse the class hierarchy
to pick up other resources. This should be less brittle than making it
necessary to update the dependencies of every subclass of a device type
whenever its external dependencies are changed.