Add facilities for tracking whether a device and its parents can be setup.
Related: #12

This requires tracking whether the format's of the device and its parents
can be setup. In some cases, the requirements on setting up certain formats
can be omitted, because setup omits these formats as well. Essentially,
wherever setup or setupParents is overridden in subclasses of Device,
the new methods must be overrridden as well.

Signed-off-by: mulhern <amulhern@redhat.com>