Treat a missing plugin as an OK situation.
Related: #12

It is best to have as many plugins as there are available, so load all that
we can. But if they are not available, just keep going.

Change "required" to "requested", since these plugins are no longer all
required.

The *_init methods will raise a GlibError if they can't load
all the requested plugins. Catch and ignore these errors,
but check what plugins are loaded and report any that are missing when done.

Signed-off-by: mulhern <amulhern@redhat.com>