Consolidate common code in DeviceFormat class methods.
This mirrors what was done long ago for the Device classes. The intention
is to improve consistency and reduce code duplication.

create and destroy are just split into pre/body/post with no way to bail.

setup and teardown pre methods return a boolean telling whether or not to
proceed with the operation.