Add ability to unmount specific mountpoints
The dynamic mountpoint detection patch revealed a problem with multiple
mountpoints. Sometimes things are mounted in multiple places and the
first one listed by the system isn't always the one you want to unmount.

This adds the mountpoint argument to unmount so that a specific location
can be selected. It must exist or else FSError will be raised.

The old code worked because it replaced the _mountpoint value so it
would unmount the last mounted location, but that wasn't correct because
you then lost the information about previous mountpoints.

With the current code the caller is expected to keep track of
mountpoints if more than one is used and to pass mountpoint= to unmount.