Fall back on mdadm info if udev info is missing for the array
If udev didn't supply the UUID of the array the device belongs to,
which means it won't be found in the format's mdUuid, assume that
mdadm got the correct UUID and use it and also assign it to the format's
mdUuid.

Otherwise, we may end up creating a new device when one is already in
the device tree.

Signed-off-by: mulhern <amulhern@redhat.com>
(cherry picked from commit a6a8ce6120ea7bd214ebfe75a1a1a30ea989cf1d)

Related: rhbz#1246003