Modify methods unavailable(), unready(), unable(), impossible().
Related: #12

Change their signature so that they return a list of errors, an empty
list, which can be interpreted as False, if there are no errors.

Change their names, so that they indicate that they return errors,
so unavailable() goes to unavailabilityErrors(), unready goes to
readinessErrors(), etc. Since impossible()/possibilityErrors() returns a
list of errors instead of a string, convert it to a string before passing
as argument to error constructor.

Signed-off-by: mulhern <amulhern@redhat.com>