Manage backing store more indepedently in loop backed test cases.
* Make the backing store before starting to use the loop devices so there
is less elapsed time between finding a free loop device and actually using
it. Get rid of it if there is a failure in locating a free loop device
or using it.

* Make self._loopMap a list of pairs, since we never do random access
anyway.

* Make the names of our temp files that back the loop devices less likely
to be reused by adding a random string to their names.

* Don't give up on deleting all loop devices the first time one loop
device is not succesfully removed.

* Give every loop device a minimum of three deletion attempts
before teardown gives up. This is apallingly ad-hoc, but almost works.
As many deletion attempts as needed is sometimes infinitely too many.

Signed-off-by: mulhern <amulhern@redhat.com>