Fix a few miscellaneous pylint-caught problems.
(1) Replace a use of filter with a list comprehension, because the former was
resulting in some weird errors from pylint.

(2) We're using print as a function everywhere else in makebumpver, so do the
same when we just want to print out a newline.

(3) raw_input in python3 has been renamed to input, so use six to figure that
out for us.