Cache mountinfo information during duration of _getActiveMounts call.
It is possible that /proc/self/mountinfo changes during execution of
_getActiveMounts, but it is also possible that /proc/mounts changes,
and we do not guard against that.

The motivation behind this is that it is expensive to parse on average half
the file for some supplementary information about btrfs everytime it is
needed.

The downside of this cacheing is that a data structure gets held in memory
for the duration of _getActiveMounts execution, although this only happens
if btrfs info is requested.

Also, there is a slight improvement in the way the slightly variable format
data in mountinfo is parsed.

Signed-off-by: mulhern <amulhern@redhat.com>