Cache data obtained from /proc/self/mountinfo.
This avoids reading and parsing the file whenever additional
btrfs information needs to be looked up.

Changes are:
* Addition of a small, special purpose cache for btrfs information
obtained from mountinfo. The only facility the cache offers is laziness.
Since it currently is instantiated and destroyed in a single method where
it is used in a very simple way it does not require other management
facilities.
* Slightly improved parsing of /proc/self/mountinfo which takes into
account some variable number fields.
* Make sure all components of key values are str or NoneType when
inserted into mounts cache or when looked up.
* Use context manager for "/proc/mounts" processing.
* Log an error if a line in "/proc/mounts" is skipped because no
subvolspec was obtained for the btrfs volume entry.
* Use named constant MAIN_VOLUME_ID instead of 5 where appropriate.

Signed-off-by: mulhern <amulhern@redhat.com>