Use python-hawkey instead of rpm-python.
The advantages are that the hawkey API is consistent between python
versions, and that it is a bit cleaner for this purpose.

python-hawkey adds two new dependencies above what rpm-python adds, but it
is required by dnf anyway, so it will likely be installed by most users.

Now raise an AvailabilityError if the number of packages is not exactly
equal to 1. This is more correct and is not going to cause more stack
traces, because the error is caught in the method that calls
packageVersion.

Now catch a possible IOError and reraise as AvailabilityError when loading
repo.

Signed-off-by: mulhern <amulhern@redhat.com>