Fix a bug in reading a size specification with a radix in the numeric part.
Since the | operator is not greedy, the match would return the first of the
three base alternatives that worked, and would split, e.g., "0.0" into
"0" and ".0" instead of "0.0" and nothing.

Now, base is a single alternative, that should capture numbers with a
decimal point correctly.

Fixes a bug introduced in commit 65703bfc851f6d919f8e66e1a635ace5835ea6c1.

Changes to the regular expression are:
* No longer allows spaces between optional sign and number. Decimal
constructor will raise exception on '- 2', so might as well not go to extra
trouble to accept it in regular expression.
* Uses verbose regular expression, to make it a little easier to read.
Uses separate compile before match call for the same reason.
* No longer allows trailing non-whitespace characters after a separating
space, e.g., "1 KiB or maybe 2". Still  allows trailing or leading
whitespace characters. Never allowed leading non-whitespace characters,
e.g., "nearly 1 PiB".
* The expression that matches the base is a lot simpler.

There is also a huge comment above the regular expression. Hopefully, it
will prevent people from making unnecessarily complicated and possibly
buggy regular expressions in future.

The only change to the behavior of the method is that the bug is fixed.

Adds a few more tests.

Signed-off-by: mulhern <amulhern@redhat.com>