Get rid of pointless test case about arguments for labeling apps.
Related: #130

It breaks when not all supporting apps are available with a KeyError.

It used to be kind of useful to make sure that the flags for the
different filesystem apps hadn't been jumbled around. It was
probably kind of useful to detect regressions.

But now that tasks branch is committed, it serves no useful purpose.
The tasks for fswritelabel module are a whole lot clearer than the tests.
And the other tests for the formats do a better job of checking for
correctness.

Signed-off-by: mulhern <amulhern@redhat.com>