Add BTRFSValueError error and use in btrfs related code (#1202877)
Related: rhbz#1019685
Related: rhbz#1202877

(cherry picked from commit d11015c37f20fbcc25ec28cb48b50542d8e28141)

From BTRFS related methods that have ValueError-like exceptions raise
BTRFSValueError instead of ValueError. Note that this will not change the
behavior of code that calls these method and catches a ValueError since
BTRFSValueError extends ValueError. It can change the behavior of calling
code where a BTRFSError is expected; since BTRFSValueError is also a subtype
of BTRFSError the formerly ValueError exception will be caught where
previously it would not. There are no instances like this in anaconda.
There are two in blivet, in devices.py, and in both cases, this change
seems to be an improvement.

Signed-off-by: mulhern <amulhern@redhat.com>