Use partially corrupt gpt disklabels.
Tell pyparted to answer "yes" when libparted asks if we want to proceed
using the secondary gpt disklabel in the event of a corrupt primary (or
use the primary in spite of a corrupt secondary label). We only do this
while populating the devicetree.

We only do this if udev/blkid has identified the disk as containing a
disklabel. This means we are going with the system's consensus as to
which formatting to acknowledge in the event there are multiple
signatures on the disk.

Based on commit e8b5e4b0058946b77ac472b466087196f4c6810b

Resolves: rhbz#1188163