Align free regions before choosing one.
By aligning the start sector of free regions we improve our ability to
select an appropriate free region in which to place a new partition.
When the start sector alignment is not taken into consideration the
possibility arises that aligning the start sector will shrink the size
of the free region below the requested size of the partition.

Related: rhbz#1181494