Handle formatting after adding devices from format handlers.
If we don't do it right away the device could be looked up successfully
and cause confusion due to inaccurate(missing) format data.

This is only needed in the places where we add a new device from a
format handler. The device additions from add*Device are not vulnerable
to this issue.

Related: rhbz#1192004
Related: rhbz#1197582