Add support for creation of cached LVs
Determining parameters of the cache automagically is hard, but if users tell us
they want (some of) their LVs cached with all the required parameters for the
cache, we can quite easily create cached LVs instead of plain ones. Doing so
requires us to change our calculations a bit, but not in any hard or too
invasive way.

Related: rhbz#1120421