Close the unused half of pipes after the fork
Closing the writing end of the pipe in the parent is important to ensure
that errors in the child that close the pipe result in the pipe being
fully closed.  Otherwise an error in the child will cause the parent to
hang reading from the pipe instead of receiving an EOFError.

Closing the reading end of the pipe in the child is less important but
ensures that bugs in libiscsi can't do anything dumb to the file
descriptor.