Tear down all devices after finding existing installations
When trying to find existing installations we setup and mount devices (to check
if they have /etc/fstab). Setting a device up results in its parents being setup
too (recursively) so when done we need to tear all/rest of the devices down so
that they are not left activated as that could cause problems later (as for
example in the bug where a thin pool was activated/setup on setup of a thin LV
and never deactivated/torn down).

Resolves: rhbz#1182229