use floor division in calculating amount to reclaim (#1245446)
Using true division results in a float, which eventually causes
a crash when the value is passed to sectorsToSize(), which
tries to multiply it by a Size. Multiplying a float by a Size
doesn't work. sectorsToSize() does specify that it needs an int.
This is a python2 vs. python3 thing: in py2, int/int returns
an int. See PEP 238.
This fixes #1245854 and #1245446 (tested).