Use slow as well as fast PVs for cached LV's non-cache part
The requested size for the LV's non-cache part may be bigger than the free space
on the slow PVs while it can safely fit in the set of slow and fast PVs together
with the cache.

An example of this is when a kickstart is used with two PVs in a VG and LV(s)
with --cachesize and --grow. In that case the LV is grown to the maximum size
filling up the VG together with the cache and other LVs (and their caches).