Fix some logical problems in write_dasd_conf (#1248949)
First, don't continue to attempt writing /etc/dasd.conf if users are
not on s390x _and_ there were no dasds used during installation. Not
'or'.

Second, self.dasd in blivet is empty for some reason(?), but
devicetree.dasd contains the list of DASDs used during installation,
so pass this in to write_dasd_conf.

Resolves: rhbz#1248949