Implement metadata size reporting for the LVM cache
Metadata size is a useful information when doing calculations with LVs' sizes
for example when growing LVs in a VG. We need to report the actual (if the cache
exists) or planned (desired or default for a non-existing cache) metadata
size. However, instead of complicating many places in the codebase by adding
extra metadata space for requested caches, we just subtract the default metadata
size from the requested size of the cache so that the whole cache
(data+metadata) fits into the space or requested size. If both size and metadata
size are specified, we respect them.

Please note that this adds the metadata size parameter only to the internal
LVMCache class. A full support of letting the user code to set metadata size
would/will require more changes in multiple places. This is however definitely
the compulsory first step.