Add a property for Requests to reserve some extra space
This is generic concept for requests that require some extra space that
shouldn't be included in their base size because it is extra/external space
taken from the same Chunk. The only usage/case right now is for LVRequests of
cached LVs to reserve space for their caches in a nicer and cleaner way, but
more are expected to come in the future because many devices require some fixed
extra space that is not part of their real/usable size (usually some metadata
space).