Don't store UUIDs or labels of multipath members.
multipath itself has absolutely no identifying metadata, so whatever
metadata found on the member is related to the multipath set's
formatting -- not the the multipath set itself (as is normally the case
when metadata is associated with a member of a container/aggregate
device). The same metadata should appear in the udev entry for the
multipath, so we can (and should) ignore it in the members' entries.

(cherry picked from commit d478f1fb4d68a3a342bed28b43ff34131f02850d)

Resolves: rhbz#1254232