Do not reserve extra space for metadata in a VG with RAID PVs
Reserving 5 extents per disk per LV for metadata if a VG is on top of RAID PVs
is a weird magic that saved us troubles with placing LVM on top of RAID, but
even at the first glance it's just wrong and doesn't make sense. LVM doesn't
allocate any extra meta data if an LV is added to a VG (unless it is a meta data
LV of course) and the RAID device below it (the PV) is just a block device with
some size and no matter how many disks are below, the meta data space used by
LVM is just subtracted from the RAID device's (PV's) size. Moreover the LVM
extent size has nothing to do with the RAID device so it's like mixing apples
and oranges.

If things don't fit we need to tweak our calculations of free space/size in/of
an MD RAID device (depending on the level and underlying disks) since that's the
only place where these tweaks really make sense. However, after quite a lot of
testing it seems that we are reserving more than enough space for the meta data.

(cherry picked from commit cfad323514a42fffccc54a60c92fede4c8368679)

Resolves: rhbz#1284660

Conflicts:
    blivet/devices/lvm.py
    blivet/partitioning.py