Calculate the MD RAID superblock size from the right size
libblockdev inherited from blivet a very precise algorithm to calculate MD RAID
superblock (meta data) size which was taken from the mdadm's sources. However,
in mdadm, this algorithm is applied to a size of a single member device not the
size of the whole RAID. Which makes sense, because the superblock area is on
every member and it contains meta data about that particular member. Thus we
need to pass the size of the smallest member device to the algorithm calculating
the superblock size, not the total size of the whole RAID.

Related: rhbz#1284660

(cherry picked from commit 716d9627e08912a993f759601f3c70b89cd082bf)