Account for bigger LVM meta data due to alignment on MD RAID
As explained in the inline comment LVM aligns data and meta data according to
the underlying block device. In case of a MD RAID device where typically a
512KiB chunks are used, this may result in twice as big space allocated for meta
data as if a plain partition was used. Since the default size is 1 MiB, we can
safely just use a double size in case an MD RAID device is used as a PV because
we can be at most 1 MiB and thus one extent off in the end which is okay.