Duplicate VG names are problem even if their disks are ignored (#1198367)
Even if user tells us we should ignore a disk that contains a VG with the same
name as a VG on some not-ignored disk, it is a problem, because LVM has no idea
about our ignored disks and thus fails to do things we need to do. More
precisely, we tell LVM to filter the ignored disk out, but that only leads to it
seeing a VG with no existing or not-ignored PVs which is no-GO for it.

Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>