Make sure devices are torn down in findExistingInstallations (#1261439)
It's not that uncommon that findExistingInstallations() raises an
exception. Even in such case we need to tear all devices down if requested. Also
such exceptions should be just logged and not propagated as this function is a
"best-effort" one.

(port of commits bbe124c94772cb3313a05e9a00a663f75084be55 and
71f77ec0e7f657a3e9bffd03b859e845b2598eb7 from rhel7-branch merged together)