Try to do fsck if resize fails before giving up (#1251396)
We may need to do fsck in many steps that go one after another -- getting FS
info, getting FS' minimum size, resize. It's usually not needed to run the fsck
multiple times so we could, theoretically, hold and check some "fsck_ran" flag,
but would it always work? A more robust approach seems to be to try to run the
FS operation and if it fails, try once more this time with a check first. We
already do this when getting FS info and minimum size so let's do the same when
doing resize.