Don't teardown FSs when searching for installed systems (#1252902)
When looking for existing installations the FSs are mounted to the
sysroot but if the device is protected it can't be teared down.

Umount the disks instead of teardown. The teardownAll() method is
used when all devices were tested.

Related: rhbz#1252902
(cherry picked from commit 05feef37e2c5ae8fd63abe2198ccd8d55e09b08a)