Wait for auto-activation of LVs when lvmetad is running. (#1261621)
When lvmetad is running, activating the last PV in a VG will trigger
automatic activation of the LVs. This happens asynchronously, however,
so we have to just wait for it to be done. Since it is possible to
configure which VGs/LVs get auto-activated, we only wait for 30
seconds. After that, we will try to activate the LV ourselves.

(cherry picked from commit 6fc20d30bab3755e626d2ff02c7f3c77b923b170)

Resolves: rhbz#1261621