warn on failure to scan partitions
I had a Fedora install fail due to https://bugzilla.redhat.com/1263832,
and I wonder if we should be treating failures at this point in the code
as warnings rather than exceptions.  This would permit blivet to continue
scanning other devices.

This would seem to be in line with the comments at the beginning of this
section of code, which suggest that the desired behavior is to ignore
the device rather than throwing an exception:

    # Ignore partitions on:
    #  - devices we do not support partitioning of, like logical volumes
    #  - devices that do not have a usable disklabel
    #  - devices that contain disklabels made by isohybrid

This commit causes blivet to log and error and return if it is unable to
scan device partitions, rather than raising an exception.