Add a 'debug' flag and use it for LVM debugging
Many times we have been asked by the LVM team to use the '-vvvv' option whenever
we do any LVM action so that the logs contain more (debugging)
information. However, getting the output of every single LVM command cluttered
with a lot of debugging info would be confusing for both us reading the logs and
the code parsing the output of such commands. The trade-off solution is to pass
the 'log{level=7}' config with all our commands that causes LVM to put the extra
info from '-vvvv' to a log. The default is syslog (i.e. journal), but since LVM
produces a huge amount of debugging log messages with level=7, it's better to
put it them into a separate file.