edd: Try to make EDD support actually match what the kernel does.
Basically the parser for ATA will kind of work with SCSI if lun==0, but
it won't work with SATA or ATA at all, and the parser we had for SCSI
will only work with virtio.

This patch separates those all out by which kernel driver is actually
providing the sysfs devices, and tries to match the EDD fields up with
how they're actually provided by the kernel.

Signed-off-by: Peter Jones <pjones@redhat.com>