Add a udev settle call after instantiating parted.Disk.
udev sees parted close a rw fd to the disk (in this case an md fwraid)
and tries to be helpful by telling the kernel to re-read the partition
table. That causes the partitions to disappear briefly, which can be
problematic.

(cherry picked from commit 5724d17e655f87479ce741296551a72d11901cf9)

Resolves: rhbz#1267858