edd: Add missing directories for absurd_virt test case.
Some of the directories from f5e4a2bced8d didn't get added, because "git
add" silently accepts empty directories without bothing to tell you it's
not going to do anything with them.  But in this case, the directory
path itself is important.

So add some stub files in them.

Signed-off-by: Peter Jones <pjones@redhat.com>