Make distutils.filelist.findall() do the right thing with symlinks maybe.
I'm not sure this would be right if we actually wanted to /include/ the
symlink itself, but since so far they're all in directories we don't
include in the tarball, this will at least make the traversal of them
not fail.

This fixes "make local" with the EDD test case data.

Signed-off-by: Peter Jones <pjones@redhat.com>