Check if device is a DASDDevice in make_unformatted_dasd list. (#1273009)
This is more properly fixed in anaconda, but for f23-branch it's
easier and simpler to "fix" it here in blivet. Since anaconda is
passing in the entire devicetree (due to another bug I fixed, #
escapes me right now), it's simple enough to verify that the
device we're operating on is a DASD before checking any DASD
specific attributes.

Resolves: rhbz#1273009