Let LVMLogicalVolumeDevice specify PVs it should be allocated from
This allows for more precise configuration of how the LVM setup should look
like. We already have some rudimentary support for specifying PVs for caches
which is a must, but it's useful for all LVs in general.

Thanks dlehman@redhat.com for pieces of code and ideas for this patch!