Add base classes for type-specific Populator helpers.
The main components of each class are the 'match' class method and
the 'run' method. The 'match' method is used to determine which
helper class is appropriate for the given udev data set. The 'run'
method is what instantiates the StorageDevice or DeviceFormat
instance and updates the DeviceTree as necessary.